+++ crates/sncast/src/main.rs
@@ -10,10 +10,10 @@ use sncast::response::print::{print_command_result, OutputFormat};
 use camino::Utf8PathBuf;
 use clap::{Parser, Subcommand};
 use sncast::helpers::constants::{DEFAULT_ACCOUNTS_FILE, DEFAULT_MULTICALL_CONTENTS};
-use sncast::helpers::scarb_utils::{build, get_scarb_manifest, parse_scarb_config, CastConfig};
+use sncast::helpers::scarb_utils::{build, get_scarb_manifest, parse_scarb_config, BuildConfig, CastConfig, get_scarb_metadata_with_deps, get_package_metadata};
 use sncast::{
     chain_id_to_network_name, get_account, get_block_id, get_chain_id, get_nonce, get_provider,
-    BuildConfig, NumbersFormat, WaitForTx,
+    NumbersFormat, WaitForTx,
 };
 use starknet::providers::jsonrpc::HttpTransport;
 use starknet::providers::JsonRpcClient;
@@ -121,9 +121,21 @@ fn main() -> Result<()> {
     let runtime = Runtime::new().expect("Failed to instantiate Runtime");
 
     if let Commands::Script(script) = cli.command {
+        let manifest_path = match cli.path_to_scarb_toml {
+            Some(path) => path,
+            None => get_scarb_manifest().context("Failed to obtain manifest path from Scarb")?,
+        };
+        let metadata = get_scarb_metadata_with_deps(&manifest_path)?;
+        let package_metadata = get_package_metadata(&metadata, &manifest_path)?;
+        let mut artifacts = build(&BuildConfig {
+            scarb_toml_path: manifest_path.clone(),
+            json: cli.json,
+        }).expect("Failed to build script");
         let mut result = starknet_commands::script::run(
             &script.script_module_name,
-            &cli.path_to_scarb_toml,
+            &metadata,
+            &package_metadata,
+            &mut artifacts,
             &provider,
             runtime,
             &config,
@@ -170,11 +182,10 @@ async fn run_async_command(
                     get_scarb_manifest().context("Failed to obtain manifest path from Scarb")?
                 }
             };
-            let build_config = BuildConfig {
+            let artifacts = build(&BuildConfig {
                 scarb_toml_path: manifest_path,
                 json: cli.json,
-            };
-            let artifacts = build(&build_config)?;
+            })?;
             let mut result = starknet_commands::declare::declare(
                 &declare.contract,
                 declare.max_fee,
