+++ crates/sncast/src/starknet_commands/script.rs
@@ -38,11 +38,10 @@ use scarb_api::{package_matches_version_requirement, StarknetContractArtifacts};
 use scarb_metadata::{Metadata, PackageMetadata};
 use semver::{Comparator, Op, Version, VersionReq};
 use sncast::helpers::scarb_utils::{
-    build, get_package_metadata, get_scarb_manifest, get_scarb_metadata_with_deps, CastConfig,
+    build, get_package_metadata, get_scarb_manifest, get_scarb_metadata_with_deps, BuildConfig, CastConfig,
 };
 use sncast::response::print::print_as_warning;
 use sncast::response::structs::ScriptResponse;
-use sncast::BuildConfig;
 use starknet::accounts::Account;
 use starknet::core::types::{BlockId, BlockTag::Pending, FieldElement};
 use starknet::providers::jsonrpc::HttpTransport;
@@ -253,29 +252,18 @@ impl<'a> ExtensionLogic for CastScriptExtension<'a> {
 
 pub fn run(
     module_name: &str,
-    path_to_scarb_toml: &Option<Utf8PathBuf>,
+    metadata: &Metadata,
+    package_metadata: &PackageMetadata,
+    artifacts: &mut HashMap<String, StarknetContractArtifacts>,
     provider: &JsonRpcClient<HttpTransport>,
     tokio_runtime: Runtime,
     config: &CastConfig,
 ) -> Result<ScriptResponse> {
-    let manifest_path = match path_to_scarb_toml.clone() {
-        Some(path) => path,
-        None => get_scarb_manifest().context("Failed to obtain manifest path from Scarb")?,
-    };
-    let build_config = BuildConfig {
-        // todo: to main?
-        scarb_toml_path: manifest_path.clone(),
-        json: false,
-    };
-    let metadata = get_scarb_metadata_with_deps(&manifest_path)?;
     warn_if_sncast_std_not_compatible(&metadata)?;
-    let mut artifacts = build(&build_config).expect("Failed to build script");
-    let package_metadata = get_package_metadata(&metadata, &manifest_path)?;
+    ensure_lib_artifact(&metadata, package_metadata, artifacts)?;
 
-    ensure_lib_artifact(&metadata, package_metadata, &mut artifacts)?;
-
-    let artifact = artifacts.get("__sncast_lib_contract").ok_or(anyhow!("Failed to find artifacts in starknet_artifacts.json file. Make sure you have enabled sierra and casm code generation in Scarb.toml"))?;
-    //let path = compile_script(path_to_scarb_toml.clone())?;
+    let artifact = artifacts.get("__sncast_script_lib")
+        .ok_or(anyhow!("Failed to find script artifact."))?;
 
     let sierra_program = serde_json::from_str::<VersionedProgram>(&artifact.sierra)
         .with_context(|| "Failed to deserialize Sierra program")?
@@ -416,47 +404,9 @@ fn ensure_lib_artifact(
         casm: String::new(),
     };
 
-    artifacts.insert("__sncast_lib_contract".to_string(), lib_artifacts);
+    artifacts.insert("__sncast_script_lib".to_string(), lib_artifacts);
     Ok(artifacts.clone())
 }
-
-// fn compile_script(path_to_scarb_toml: Option<Utf8PathBuf>) -> Result<Utf8PathBuf> {
-//     // let scripts_manifest_path = path_to_scarb_toml.unwrap_or_else(|| {
-//     //     get_scarb_manifest()
-//     //         .context("Failed to retrieve manifest path from scarb")
-//     //         .unwrap()
-//     // });
-//     // ensure!(
-//     //     scripts_manifest_path.exists(),
-//     //     "The path = {scripts_manifest_path} does not exist"
-//     // );
-//
-//     // ScarbCommand::new_with_stdio()
-//     //     .arg("build")
-//     //     .manifest_path(&scripts_manifest_path)
-//     //     .run()
-//     //     .context("failed to compile script with scarb")?;
-//
-//     let metadata = get_scarb_metadata_with_deps(&scripts_manifest_path)?;
-//
-//     warn_if_sncast_std_not_compatible(&metadata)?;
-//
-//     let package_metadata = get_package_metadata(&metadata, &scripts_manifest_path)?;
-//
-//     let filename = format!("{}.sierra.json", package_metadata.name);
-//     let path = metadata
-//         .target_dir
-//         .unwrap_or(metadata.workspace.root.join("target"))
-//         .join(metadata.current_profile)
-//         .join(filename.clone());
-//
-//     ensure!(
-//         path.exists(),
-//         "The package has not been compiled, the file at path = {path} does not exist"
-//     );
-//
-//     Ok(path)
-// }
 
 // taken from starknet-foundry/crates/forge/src/test_case_summary.rs
 /// Helper function to build `readable_text` from a run data.
